net-misc/dhcpcd: delete dhcpcd
As Flatcar relies on systemd-networkd for network configurations,
it is not needed to keep dhcpcd in production images at all.

According to the commit
https://github.com/kinvolk/coreos-overlay/commit/9be90f06e838 ,
it was added back in 2014 just because systemd-networkd was not mature
enough. That was already ~7 years ago, so we can safely assume that
the issue had been already gone, so we can simply use systemd-networkd.