github: Limit the depth of Gentoo checkout
It is quite a bit of data to download for no real reason. We are
trying to update packages here, so we will be grabbing them from the
most recent commit that made the changes to the package. With the
advancement the package updates effort, we possibly can later lower
the number of the fetched commits even further.